refactor- 초기 데이터 클래스 위치 변경, PostConstruct -> CommandLineRunner로 방식 변경
