+++ backend/src/main/java/com/example/nbe341team02/domain/delivery/data/DeliveryPolicyInitData.java
@@ -1,20 +1,20 @@
-package com.example.nbe341team02.global.initData.delivery;
+package com.example.nbe341team02.domain.delivery.data;
 
 import com.example.nbe341team02.domain.delivery.entity.DeliveryTimePolicy;
 import com.example.nbe341team02.domain.delivery.repository.DeliveryTimePolicyRepository;
-import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalTime;
 
 @Component("deliveryTimePolicyInitData")
 @RequiredArgsConstructor
-public class DeliveryTimePolicyInitData {
+public class DeliveryPolicyInitData implements CommandLineRunner {
     private final DeliveryTimePolicyRepository deliveryTimePolicyRepository;
 
-    @Value("${default.delivery.hour: 0}")
+    @Value("${default.delivery.hour: 14}")
     private int deliveryTimeHour;
 
     @Value("${default.delivery.minute: 0}")
@@ -23,8 +23,8 @@ public class DeliveryTimePolicyInitData {
     @Value("${default.delivery.second: 0}")
     private int deliveryTimeSecond;
 
-    @PostConstruct
-    public void init() {
+    @Override
+    public void run(String... args) throws Exception {
         DeliveryTimePolicy deliveryTimePolicy = DeliveryTimePolicy.builder()
                 .deliveryTime(LocalTime.of(deliveryTimeHour, deliveryTimeMinute, deliveryTimeSecond))
                 .build();
